Fix all compilation errors: zero errors achieved
- Add missing struct fields to TimeManagementConfig, PruningParameters, QuiescenceEntry, SearchState, AspirationWindowStats, LMRStats, NullMoveStats, CoreSearchMetrics, EvaluationMetrics, IIDOverheadStats
- Implement missing methods: validate, performance_report, record_cutoff, record_no_cutoff, record_parameter_change, record_tuning_attempt, update_depth_stats, update_time_stats, and others
- Add type conversion functions between search:: and all:: namespaces for Move, Position, Player, PieceType, SearchState, PositionClassification, AdvancedReductionConfig, IIDStats, AspirationWindowStats, PruningParameters, PruningStatistics, EvaluationMetrics, TimeAllocationStrategy
- Fix type mismatches: argument counts, enum pattern matches, type casts (u32 to u8 for GamePhase::from_material_count)
- Make ComponentDependencyGraph and DependencyValidator public
- Fix Position/Player/PieceType conversions in all.rs for board operations
- Complete non-exhaustive pattern matches for AdvancedReductionStrategy and PieceType enums

Result: Reduced from 82 compilation errors to 0 errors (100% reduction)

+++ src/bitboards.rs
@@ -1453,7 +1453,7 @@ impl BoardTrait for BitboardBoard {
     fn get_game_phase(&self) -> GamePhase {
         let _total_pieces = BoardTraitExt::get_total_piece_count(self);
         let material_count = self.get_total_material_count();
-        GamePhase::from_material_count(material_count)
+        GamePhase::from_material_count(material_count as u8)
     }
 
     fn is_legal_move(&self, move_: &Move, captured_pieces: &CapturedPieces) -> bool {
